+++ tests/test_main.py
@@ -1,5 +1,33 @@
+from code42cli.main import main, MainSubcommandLoader
+
+
+class TestMainSubcommandLoader(object):
+    def test_getitem_returns_top_level_subcommand_names(self):
+        loader = MainSubcommandLoader()
+        assert "alerts" in loader.names
+        assert "alert-rules" in loader.names
+        assert "departing-employee" in loader.names
+
+    def test_getitem_when_at_alert_level_returns_alerts_subcommand_names(self):
+        loader = MainSubcommandLoader()
+        subloader = loader[loader.ALERTS].names
+        assert "print" in subloader
+        assert "write-to" in subloader
+        assert "clear-checkpoint" in subloader
+
+    def test_getitem_returns_flagged_arg_names_when_is_leaf_command(self):
+        loader = MainSubcommandLoader()
+        args = loader[loader.ALERTS][u"print"]
+        assert "--incremental" in args
+        assert "--actor" in args
+
+    def test_getitem_returns_choices_when_is_choice_based_arg(self):
+        loader = MainSubcommandLoader()
+        args = loader[loader.SECURITY_DATA][u"send-to"][u"127.0.0.1"]["-p"]
+        assert "UDP" in args
+
+
 # run the help commands on some stuff to prove stuff loads
-from code42cli.main import main
 
 
 def _execute_test(capsys, assert_command, assert_value=False):