+++ CHANGELOG.md
@@ -8,11 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
-## 0.7.1 - 2020-06-11
+## 0.7.2 - 2020-06-11
 
 ### Fixed
 
 - Fixed bug that caused `alert-rules list` to error due to page size restrictions on backing service.
+
+## 0.7.1 - 2020-06-10
+
+### Fixed
+
+- Issue that prevented alerts from being retrieved successfully via `code42 alerts` commands due to a change in its backing API.
 
 ## 0.7.0 - 2020-06-08
 
@@ -83,6 +89,8 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 - A progress bar that displays during bulk commands.
 
 - Short option `-u` added for `code42 high-risk-employee add-risk-tags` and `remove-risk-tags`.
+
+- Tab completion for bash and zsh for Unix based machines.
 
 ### Fixed